pyproject.toml
[tool.poetry]
name = "gamja-be"
version = "0.1.0"
description = ""
authors = ["dayeonkimm <ekdyd516@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
drf-spectacular = "^0.27.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
pytz = "^2024.2"
cryptography = "^43.0.3"
django-cryptography = "^1.1"
celery = "^5.4.0"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
